+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 11.0
 - PostgreSQL rename "rank" to "class_rank" in set_class_rank_mp() for consistency with MySQL in rosariosis.sql
 - HTML put "" around the link href if no spaces in $link & no other attributes in Buttons.php & program wide
 - HTML5 CSS use details + summary instead of .toggle hack in index.php, stylesheet.css & colors.css
+- SQL add TITLE column in GetTeacher.fnc.php
 
 Changes in 10.9.5
 -----------------